+++ apps/prairielearn/src/ee/routers/lti13.ts
@@ -8,23 +8,11 @@ import authnMiddleware from '../../middlewares/authn.js';
 import csrfToken from '../../middlewares/csrfToken.js';
 import lti13Auth from '../auth/lti13/lti13Auth.js';
 import { getInstitutionAuthenticationProviders } from '../lib/institution.js';
+import { configPage, jwksPage } from '../lib/lti13Metadata.js';
 import { selectLti13Instance } from '../models/lti13Instance.js';
 import lti13CourseNavigation from '../pages/lti13CourseNavigation/lti13CourseNavigation.js';
-import lti13InstancePages from '../pages/lti13Instance/lti13Instance.js';
 
 const router = Router({ mergeParams: true });
-
-const SSORequired = asyncHandler(async (req, res, next) => {
-  const lti13_instance = await selectLti13Instance(req.params.lti13_instance_id);
-  const instAuthProviders = await getInstitutionAuthenticationProviders(
-    lti13_instance.institution_id,
-  );
-
-  if (!instAuthProviders.some((a) => a.name === 'LTI 1.3')) {
-    throw new Error('Institution does not support LTI 1.3 SSO authentication');
-  }
-  next();
-});
 
 router.use(
   '/:lti13_instance_id/',
@@ -42,16 +30,31 @@ router.use(
   }),
 );
 
-router.use('/:lti13_instance_id/auth', SSORequired, lti13Auth);
+router.use('/:lti13_instance_id/config', configPage);
+router.use('/:lti13_instance_id/jwks', jwksPage);
+
+// Everything past this middleware requires LTI 1.3 SSO to be configured
+router.use(
+  '/:lti13_instance_id/',
+  asyncHandler(async (req, res, next) => {
+    const lti13_instance = await selectLti13Instance(req.params.lti13_instance_id);
+    const instAuthProviders = await getInstitutionAuthenticationProviders(
+      lti13_instance.institution_id,
+    );
+
+    if (!instAuthProviders.some((a) => a.name === 'LTI 1.3')) {
+      throw new Error('Institution does not support LTI 1.3 SSO authentication');
+    }
+    next();
+  }),
+);
+
+router.use('/:lti13_instance_id/auth', lti13Auth);
 router.use(
   '/:lti13_instance_id/course_navigation',
-  SSORequired,
   authnMiddleware, // authentication, set res.locals.authn_user
   csrfToken,
   lti13CourseNavigation,
 );
-
-// lti13InstancePages is a catch all for some small pages so put it last
-router.use('/:lti13_instance_id/', lti13InstancePages);
 
 export default router;
